Terrible attempt at adding closures
This code is a good representation of my exploratory closure attempt. Ultimately I made the code way more brittle & it doesn't actually work, but it's probably 95% there in that regard. Key take away for second attempt is to generate separate closure functions instead of trying to overload all the functionality into a normal function definition, because it breaks typing in a lot of different ways.