Added struct methods
I'm not sure that I like all the added data passing required by this current implementation (prog is now held by the compiler and unifier, because they need to know which struct has which methods, and which methods map to which functions). But it appears to work just fine. Currently, calling a struct method will require a heap allocation to create the closure, but this should be really easy to optimize later down the line (don't make a closure if the method is called immediately after it is accessed.)