Fixed type inference container bug
Containers can go in the type inference substitution map. However, the subtypes that they contain are not necessarily going to be consistent during type resolution. Each container now has an ID that identifies it, which is used for container hashing. This allows containers to be looked up by their ID rather than a hash of the struct.