Simple pipeline desugaring
Can't actually meaningfully use the output right now, because I don't have a way to dynamically grow lists or allocate a list of a fixed size. Both of these needs to be added. Went a bit down the rabbit hole of implementing struct methods on base types. That will need to be done, but isn't very straightforward at the type inference stage right now.