Added type validation stage
Basically, the type inference system infers types, but it doesn't enforce certain type rules, such as "these types are addable" or "these types can be treated as numbers". So there's now an extra stage after type inference that does that. The error messages/error handling is still terrible, but that will be handled soon.