Rewrote type inference system
The old way was too complex and very hard to reason about. I replaced it with this version which should have less edge cases and should be easier to work with. As a side effect, it appears to be much much faster. Still somewhat a work in progress, not all previously supported features are still supported.